+++ apps/astarte_data_updater_plant/lib/astarte_data_updater_plant/data_updater/server.ex
@@ -1,7 +1,7 @@
 #
 # This file is part of Astarte.
 #
-# Copyright 2017 - 2023 SECO Mind Srl
+# Copyright 2017 - 2025 SECO Mind Srl
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,16 +22,22 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
   alias Astarte.DataUpdaterPlant.DataUpdater.Impl
   alias Astarte.DataUpdaterPlant.MessageTracker
 
-  def start(realm, device_id, message_tracker, opts \\ []) do
-    GenServer.start(__MODULE__, {realm, device_id, message_tracker}, opts)
+  require Logger
+
+  def start_link({realm, device_id, _message_tracker} = args) do
+    name = {:via, Horde.Registry, {Registry.DataUpdater, {realm, device_id}}}
+    GenServer.start_link(__MODULE__, args, name: name)
   end
 
+  @impl GenServer
   def init({realm, device_id, message_tracker}) do
     timeout = Config.data_updater_deactivation_interval_ms!()
 
+    Process.flag(:trap_exit, true)
     {:ok, Impl.init_state(realm, device_id, message_tracker), timeout}
   end
 
+  @impl GenServer
   def handle_cast({:handle_connection, ip_address, message_id, timestamp}, state) do
     timeout = Config.data_updater_deactivation_interval_ms!()
 
@@ -43,6 +49,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
     end
   end
 
+  @impl GenServer
   def handle_cast({:handle_disconnection, message_id, timestamp}, state) do
     timeout = Config.data_updater_deactivation_interval_ms!()
 
@@ -55,6 +62,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
   end
 
   # TODO remove this when all heartbeats will be moved to internal
+  @impl GenServer
   def handle_cast({:handle_heartbeat, message_id, timestamp}, state) do
     if MessageTracker.can_process_message(state.message_tracker, message_id) do
       new_state = Impl.handle_heartbeat(state, message_id, timestamp)
@@ -64,6 +72,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
     end
   end
 
+  @impl GenServer
   def handle_cast({:handle_internal, payload, path, message_id, timestamp}, state) do
     if MessageTracker.can_process_message(state.message_tracker, message_id) do
       case Impl.handle_internal(state, payload, path, message_id, timestamp) do
@@ -76,6 +85,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
     end
   end
 
+  @impl GenServer
   def handle_cast({:handle_data, interface, path, payload, message_id, timestamp}, state) do
     timeout = Config.data_updater_deactivation_interval_ms!()
 
@@ -96,6 +106,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
     end
   end
 
+  @impl GenServer
   def handle_cast({:handle_introspection, payload, message_id, timestamp}, state) do
     timeout = Config.data_updater_deactivation_interval_ms!()
 
@@ -107,6 +118,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
     end
   end
 
+  @impl GenServer
   def handle_cast({:handle_control, payload, path, message_id, timestamp}, state) do
     timeout = Config.data_updater_deactivation_interval_ms!()
 
@@ -118,6 +130,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
     end
   end
 
+  @impl GenServer
   def handle_call(
         {:handle_install_volatile_trigger, object_id, object_type, parent_id, trigger_id,
          simple_trigger, trigger_target},
@@ -140,6 +153,7 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
     {:reply, return_value, new_state, timeout}
   end
 
+  @impl GenServer
   def handle_call({:handle_delete_volatile_trigger, trigger_id}, _from, state) do
     timeout = Config.data_updater_deactivation_interval_ms!()
 
@@ -148,30 +162,50 @@ defmodule Astarte.DataUpdaterPlant.DataUpdater.Server do
     {:reply, result, new_state, timeout}
   end
 
+  @impl GenServer
   def handle_call({:dump_state}, _from, state) do
     timeout = Config.data_updater_deactivation_interval_ms!()
 
     {:reply, state, state, timeout}
   end
 
+  @impl GenServer
   def handle_call({:start_device_deletion, timestamp}, _from, state) do
     {result, new_state} = Impl.start_device_deletion(state, timestamp)
     {:reply, result, new_state}
   end
 
+  @impl GenServer
   def handle_info({:DOWN, _, :process, pid, :normal}, %{message_tracker: pid} = state) do
     # This is a MessageTracker normally terminating due to deactivation
     {:noreply, state}
   end
 
+  @impl GenServer
   def handle_info({:DOWN, _, :process, _pid, :shutdown}, state) do
     {:stop, :shutdown, state}
   end
 
+  @impl GenServer
   def handle_info({:DOWN, _, :process, _pid, _reason}, state) do
     {:stop, :monitored_process_died, state}
   end
 
+  @impl GenServer
+  def handle_info(
+        {:EXIT, _pid, {:name_conflict, {_name, _value}, _registry, _winning_pid}},
+        state
+      ) do
+    _ =
+      Logger.warning(
+        "Received a :name_confict signal from the outer space, maybe a netsplit occurred? Gracefully shutting down.",
+        tag: "name_conflict"
+      )
+
+    {:stop, :normal, state}
+  end
+
+  @impl GenServer
   def handle_info(:timeout, state) do
     :ok = Impl.handle_deactivation(state)
     :ok = MessageTracker.deactivate(state.message_tracker)