+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Update the docker-compose configuration to allow both physical and virtual devices
   to connect to Astarte, provided that the devices and the host are on the same LAN.
+- Rework RPC between AppEngine and DUP using Erlang's native clustering and
+  message-passing instead of AMQP queues:
+  [#1186](https://github.com/astarte-platform/astarte/pull/1186).
+  Fix [#699](https://github.com/astarte-platform/astarte/issues/699).
+- Rework RPC between AppEngine and VerneMQ using Erlang's native clustering and
+  message-passing instead of AMQP queues:
+  [#1194](https://github.com/astarte-platform/astarte/pull/1194).
+- Rework RPC between DUP and VerneMQ using Erlang's native clustering and
+  message-passing instead of AMQP queues:
+  [#1197](https://github.com/astarte-platform/astarte/pull/1197).
 
 ## Fixed
 - [astarte_appengine_api] Correctly handle Cassandra `varchar`s.